backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
SQLAlchemy==2.0.36
Werkzeug==3.0.3
python-docx==1.1.0
python-dotenv==1.0.1
